aquicarattino/requirements.txt
wagtail>=2.7,<2.8
wagtailfontawesome>=1.1.3,<1.2
-e git://github.com/torchbox/wagtail-markdown.git#egg=wagtail-markdown
django-widget-tweaks
django-allauth
django-comments-xtd
